% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_collage.R
\name{create_collage}
\alias{create_collage}
\title{A function to create a collage of randomly colourised image}
\usage{
create_collage(pretty_image, plot_image = FALSE, filename_to_save = NULL)
}
\arguments{
\item{pretty_image}{Define a pretty image using imager package}

\item{plot_image}{if TRUE the image will be plotted to the console. Default=FALSE}

\item{filename_to_save}{Where should the image be saved? ie ./image.jpeg. default=NULL}
}
\value{
Return a beautiful collage
}
\description{
This function allows you to produce a collage of an image with random colourings and save to file
}
\examples{
library(imager)
parrots.cp <- load.example("parrots")
create_collage(parrots.cp, plot_image=TRUE, filename_to_save = '~/Desktop/jungle.jpeg')
}
\keyword{collage}
\keyword{colours}
\keyword{image}
\keyword{pretty}
\keyword{random}
