% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_collage.R
\name{randomise_colours}
\alias{randomise_colours}
\title{A function to randomise colours in an image}
\usage{
randomise_colours(pretty_image)
}
\arguments{
\item{pretty_image}{Set the pretty_image to an imager file that you want}
}
\value{
Return a colourised image
}
\description{
This function allows you to randomise RGB values in an imager image
}
\examples{
library(imager)
parrots.cp <- load.example("parrots")
randomise_colours(parrots.cp)
}
\keyword{colours}
\keyword{image}
\keyword{pretty}
\keyword{random}
